Fixed round bug
fix: Fixed the current round bug

The current round was not being updated because I forgot to increment
it. Now it's being incremented.

Also fixed the bug where current round was not part of the SELECT query
when getting this rounds pairs. I added the WHERE clause to resolve
that.